Minor Python code cleanup

+++ languages/python/activedirectory_github_sync.py
@@ -1,3 +1,6 @@
+#!/usr/bin/python
+
+from __future__ import print_function
 import argparse, json, os, sys
 
 from agithub.GitHub import GitHub
@@ -27,19 +30,26 @@ def main():
         print('User removed from GitHub org: {}'.format(gh_login), file=sys.stderr)
         del login_mapping[gh_login]
 
-    for gh_login, ad_login in login_mapping.items():
+    for gh_login, ad_login in list(login_mapping.items()):
         if ad_login and not ad_client.retrieve_from_login(ad_login):
             print('User not in AD, REMOVED: {}/{}'.format(gh_login, ad_login), file=sys.stderr)
+            if not args.dry_run:
+                ag.orgs[args.org].members[gh_login].delete()
             del members[gh_login]
 
-    with open(args.login_mapping_file, 'w') as login_mapping_file:
-        json.dump(login_mapping, login_mapping_file)
+    if args.dry_run:
+        print(json.dumps(login_mapping, sort_keys=True, indent=4))
+    else:
+        with open(args.login_mapping_file, 'w') as login_mapping_file:
+            json.dump(login_mapping, login_mapping_file, sort_keys=True, indent=4)
+
 
 def parse_args():
     parser = argparse.ArgumentParser(formatter_class=argparse.ArgumentDefaultsHelpFormatter, allow_abbrev=False)
     parser.add_argument('org', help=' ')
     parser.add_argument('login_mapping_file', help=' ')
-    parser.add_argument('--ignore_certs', action='store_true', help=' ')
+    parser.add_argument('--dry-run', action='store_true', help=' ')
+    parser.add_argument('--ignore-ldap-certs', action='store_true', help=' ')
     for var in ('GITHUB_API_TOKEN', 'LDAP_URL', 'LDAP_DC', 'LDAP_BIND_USER', 'LDAP_PASSWORD'):
         if var not in os.environ:
             raise EnvironmentError('The {} environment variable must be defined'.format(var))
@@ -51,7 +61,7 @@ class ActiveDirectoryClient:
     def __init__(self, args):
         self.ldap_dc = args.ldap_dc
         ldap.set_option(ldap.OPT_REFERRALS, 0)
-        if args.ignore_certs:
+        if args.ignore_ldap_certs:
             ldap.set_option(ldap.OPT_X_TLS_REQUIRE_CERT, ldap.OPT_X_TLS_NEVER)
         self.connection = ldap.initialize(args.ldap_url)
         self.connection.simple_bind_s('{},{}'.format(args.ldap_dc, args.ldap_bind_user), args.ldap_password)
